Seperate the tokenizer from main
Added tokenize to its own file and added some comments aswell.
